added internal optics module.
It's more-or-less a clone of some minimal operators from microlens or
pre-Profunctor lens.

If use of optics became pervasive, it would probably be
better to depend on microlens (lens compile time is too long for xmonad). But a
few locally-defined, non-exported operators is a good place to start.

added conditional dependency on contravariant.

Stack lts-12's base doesn't have Data.Functor.Contravariant, so builds were failing.