fixed a bug I introduced in `windows`, clarified names
I managed to confuse myself about the 'new set of windows' and the 'set of new
windows'. First I was setting initial properties on the full new set of windows,
then I was withdrawing windows that weren't in the set of new windows. Fixed the
first bug already, this should fix the second.